requirements.txt
app==0.0.1
click==8.0.3
Flask==1.1.2
Flask_Login==0.6.1
Flask_SQLAlchemy==2.5.1
python-dotenv==0.20.0
Werkzeug==2.0.2
